pyproject.toml
[tool.poetry]
name = "pixaris"
version = "0.1.0"
description = "Evaluation framework for image generation"
authors = ["Team TIGA"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"
google-cloud-storage = "^2.18.2"
kubernetes = "^31.0.0"
pytest = "^8.3.4"
google-cloud-aiplatform = {extras = ["tensorboard"], version = "^1.75.0"}
tensorflow = "^2.12.0"
scikit-learn = "^1.6.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"
poetry-plugin-export = "^1.8.0"
pytest = "^8.2.2"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
